Refactor, separate ships and projectiles.
Make Fighter and PlayerShip subclass a new
Ship class instead of SpaceObj, and move
some of the weapon code into Ship.

Store projectiles and ships in separate
collections so we don't need to constantly
check if objects are projectiles or ships,
and so we don't need to loop through all
ships and projectiles when we really just
want one or the other.