Stop timing calls to JPanel.repaint.
JPanel.repaint calls addDirtyRegion, which
"adds a component in the list of components
that should be refreshed. If c already has
a dirty region, the rectangle (x,y,w,h)
will be unioned with the region that should
be redrawn." So calling JPanel.repaint many
times should not cause a backlog and
there's no reason not to call it often.